+++ Sources/Swift/Core/Tools/ViewCapture/SentryUIRedactBuilder.swift
@@ -23,7 +23,9 @@ final class SentryUIRedactBuilder {
     /// This object identifier is used to identify views of this class type during the redaction process.
     /// This workaround is required because SwiftUI's List view uses this class to display the background color
     /// with it expanding way beyond the bounds of the list.
-    private static let collectionViewListLayoutSectionBackgroundColorDecorationViewClassId = "_UICollectionViewListLayoutSectionBackgroundColorDecorationView"
+    private static let collectionViewListLayoutSectionBackgroundColorDecorationViewClassId = "_UICollectionViewListLayoutSectionBackgroundColorDecorationView" // swiftlint:disable:this identifier_name
+
+    // MARK: - Properties
 
     ///This is a wrapper which marks it's direct children to be ignored
     private var ignoreContainerClassIdentifier: ObjectIdentifier?
@@ -347,6 +349,9 @@ final class SentryUIRedactBuilder {
         }
 
         if viewTypeId == Self.collectionViewListLayoutSectionBackgroundColorDecorationViewClassId {
+            // _UICollectionViewListLayoutSectionBackgroundColorDecorationView is a special case because it is used by the SwiftUI.List view to display the background color.
+            // 
+            // We need to ignore it because it causes issues in the redaction clipping process.
             return true
         }
         return false